[vpsAdminOS] add new export option root_uid to nfsd
Since all containers run within user namespace, mounted NFS filesystems
have to be shifted into the container's user namespace as well. The
issue is that NFS does not recognize the container's root user as a
privileged user. To the remote NFS server, the container's root is just
another unprivileged user.

This patch introduces a new NFS export option root_uid, which can be
used to specify what uid should be considered as privileged in addition
to the global root. The export can be configured e.g. like:

  /directory  <host>(rw,no_root_squash,no_subtree_check,root_uid=100000)

where the namespaced root maps to 100000 on the host and remote NFS
servers.