[WIP] memcg: per-memcg kswapd
There is a kswapd kernel thread for each numa node. We will add a different
kswapd for each memcg. The kswapd is sleeping in the wait queue headed at
kswapd_wait field of a kswapd_info_t descriptor. The descriptor stores
information of node or memcg and it allows the global and per-memcg background
reclaim to share common reclaim algorithms.

This patch is a work based on v3 memcg: per cgroup background reclaim series
from Ying Han:

https://www.spinics.net/lists/linux-mm/msg17251.html
Signed-off-by: Pavel Snajdr <snajpa@snajpa.net>